Assorted non-functional changes
Bench: 2098685

+++ Renegade/Board.h
@@ -11,9 +11,9 @@
 */
 
 // Magic lookup tables
-extern uint64_t GetBishopAttacks(const int square, const uint64_t occupancy);
-extern uint64_t GetRookAttacks(const int square, const uint64_t occupancy);
-extern uint64_t GetQueenAttacks(const int square, const uint64_t occupancy);
+extern uint64_t GetBishopAttacks(const uint8_t square, const uint64_t occupancy);
+extern uint64_t GetRookAttacks(const uint8_t square, const uint64_t occupancy);
+extern uint64_t GetQueenAttacks(const uint8_t square, const uint64_t occupancy);
 
 class Board
 {
