+++ Renegade/Evaluation.h
@@ -2,9 +2,9 @@
 #include "Board.h"
 #include "Move.h"
 
-extern uint64_t GetBishopAttacks(const int square, const uint64_t occupancy);
-extern uint64_t GetRookAttacks(const int square, const uint64_t occupancy);
-extern uint64_t GetQueenAttacks(const int square, const uint64_t occupancy);
+extern uint64_t GetBishopAttacks(const uint8_t square, const uint64_t occupancy);
+extern uint64_t GetRookAttacks(const uint8_t square, const uint64_t occupancy);
+extern uint64_t GetQueenAttacks(const uint8_t square, const uint64_t occupancy);
 
 struct EvaluationFeatures {
